Control panel widgets: Split out different types
Move widgets used for settings to their own module, and the same for the
waypoints table widget. This makes the long files more manageable.
